Incorporate schema validation
Before processing any custom rules, we'll run the input yaml
through a [tv4](https://www.npmjs.com/package/tv4) jsonschema validation.

The big result here will be that the linter will now complain on fields
that are not supported or not expected, which should alleviate a lot of
issues. It will also detect issues where the type of a yaml field does not match.

Details
-------

The schema is generated from the golang definitions in
[libyaml](https://github.com/replicatedhq/libyaml). We had
to make some changes to the generated schema to accomodate some of the
unmarshal logic in libyaml, e.g. the fact that a templatable bool field
like `backup.enabled` can be a string or a boolean in the yaml means we
have to set the type in JsonSchema4 to `"type": ["string", "boolean"]`.
Same is true for some numbers, etc.

It's pretty hacked but it works. Unfortunately it means we'll not really
be able to automatically port updates from libyaml over, there will be
probably be some manual work.

The schema validation is opt-in, library users can pass their own schema
in calls to `linter.lint(yaml, opts)`, similar to how custom rules
can be defined.

This includes unit tests for some small cases as well as a 600-line
yaml that exercises most of the available features for `# kind: replicated`.

Other changes
-------

- `linter.lint()` has a lot of params now, refactored into an interface.
- refactor `lint` into a class, clean up a lot of tramp data params
- Add a `linter.lintDefault()` which will use the default rules