add ability to pass project in CLI
What I Did
------------

Add `--project` flag for setting a project to validate against.

How I Did it
------------

update `validate.ts` to check for project and override rules/schema. If
`--schema` is specified, it overrides the project schema. If
`--extraRules` is specified, they are appended to the project's rules.

How to verify it
------------

    npx replicated-lint --project replicatedShip -f path/to/ship.yaml